Feat/save entry insertions on room (#11)

+++ feature/diary/src/main/java/com/emenjivar/feature/diary/screens/browse/DiaryBrowseUiState.kt
@@ -6,5 +6,5 @@ import kotlinx.coroutines.flow.StateFlow
 data class DiaryBrowseUiState(
     val entries: StateFlow<List<DiaryEntry>>,
     val navigateToNewEntry: () -> Unit,
-    val navigateToDetailEntry: (uid: Int) -> Unit
+    val navigateToDetailEntry: (uid: Long) -> Unit
 )
